Removed reset pin on backplane
The connection between the reset pin and the backplane has been removed
as it is not required. The reset pin on the backplane served no function
as it would be near impossible to use in practice and was likely to
cause issues surrounding pullup resistors if too many boards were added.
As such, the link between the backplane and the IC reset pins has been
removed, leaving only the reset button on the board.

Fixes #1

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>